+++ src/index.ts
@@ -24,6 +24,7 @@ import "./sna/ActuatorDisabler";
 import "./sna/ActuatorEnabler";
 import "./sna/ActuatorLight";
 import "./sna/ActuatorTextBar";
+import "./sna/ActuatorSignalEmitter";
 import "./sna/ActuatorSound";
 import "./sna/SensorClick";
 import "./sna/SensorContact";
@@ -41,6 +42,13 @@ declare var noEditWorlds: any;
 window.onload = main;
 
 function main() {
+
+    //adjust for hi dpi monitors
+
+    var el = document.body;
+    var fs = window.getComputedStyle(el, null).getPropertyValue('font-size');
+    var fontSize = parseFloat(fs); 
+    el.style.fontSize = (fontSize/window.devicePixelRatio) + 'px';
 
     //http://www.html5gamedevs.com/topic/35741-sceneloaderappend-do-not-load-images-of-models-correctly/?tab=comments#comment-205631
     //in 4.0.3 setting this to true broke babylon file exported from GLB